IMS_1.2(Inventory managment system)
if we use 2 if & a else, it's wronge. because machine check first if(triger if it true), then moves to 2nd IF. if it not trigger then the else got trigger. therefore we should use elif at 2nd time. machine look for first IF. if it triger machine declare other 2. if we use IF for 2nd time, machine thinks last 2 things are a one thing. so 2nd IF not get trigger,machine will ezecute else